worker.py
import syft as sy
import numpy as np

# Configuration
DOMAIN_NAME = "example_domain"

# Connect to domain
server = sy.orchestra.launch(name=DOMAIN_NAME, server_side_type="low", dev_mode=False)
client = server.login()  

# Download dataset 
datasets = client.datasets

if datasets:
    dataset = datasets[0]
    data = dataset.assets[0].data
    mock = dataset.assets[0].mock
    print("Downloaded dataset:", data)

    # Process the data (e.g., train a model)
    # Here model training code

    # For demonstration
    print("Mock data:", mock)
    print("Real data:", data)
    
    # Upload results (if applicable)
    # Implement result submission logic here

else:
    print("No datasets available")

# Cleanup
server.python_server.cleanup()
server.land()
